requirements.txt
numpy==2.2.6
pandas==2.2.3
scikit_learn==1.6.1
tensorflow_cpu==2.19.0
~atplotlib==3.8.4
